requirements/heroku.txt
-r production.txt
dj-static==0.0.6
gunicorn==19.3.0
boto==2.38.0
django-storages==1.1.8
django-toolbelt==0.0.1
